Move hook to common
